requirements.txt
pandas==1.2.5
mysql-connector-python==8.0.25
psycopg2-binary==2.9.1
requests~=2.25.1